Add settings

+++ sources/Google.Solutions.IapDesktop.Extensions.Rdp/Views/RemoteDesktop/RemoteDesktopPane.cs
@@ -180,8 +180,8 @@ namespace Google.Solutions.IapDesktop.Extensions.Rdp.Views.RemoteDesktop
                 // about twice the configured timeout before sending a 
                 // OnDisconnected event.
                 //
-                advancedSettings.singleConnectionTimeout = settings.ConnectionTimeout;
-                advancedSettings.overallConnectionTimeout = settings.ConnectionTimeout;
+                advancedSettings.singleConnectionTimeout = (int)settings.ConnectionTimeout;
+                advancedSettings.overallConnectionTimeout = (int)settings.ConnectionTimeout;
 
                 //
                 // Behavior settings.
